% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeNET_DGN-extract.R,
%   R/DataGeNET_RDF-extract.R
\name{extract}
\alias{extract}
\alias{DataGeNET.DGN-methods}
\alias{DataGeNET.Dis-methods}
\title{Obtain the raw query from DisGeNET from a \code{DataGeNET.DGN} object.}
\usage{
\S4method{extract}{DataGeNET.DGN}(object)

\S4method{extract}{DataGeNET.RDF}(object)
}
\arguments{
\item{object}{Object of class \code{DataGeNET.RDF}}
}
\value{
A \code{data.frame} containing the raw result from DisGeNET

A \code{data.frame} containing the raw result from DisGeNET
}
\description{
Obtain the raw query from DisGeNET from a \code{DataGeNET.DGN} object.

Obtain the raw query from DisGeNET from a \code{DataGeNET.RDF} object.
}
\examples{
\dontrun{
#Being x an DataGeNET.DGN
qr <- extract(x) # Get number of unique diseases
}
\dontrun{
#Being x an DataGeNET.RDF
qr <- extract(x) 
}
}
