% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene2indication.R
\name{gene2indication}
\alias{gene2indication}
\title{Query for given gene(s) and generates an \code{DataGeNET.RDF}
Given the NCBI gene identifier, retrieves the information
related to drugs which interacts with the gene product and
their indications and creates an object of type \code{DataGeNET.RDF}.}
\usage{
gene2indication(gene, verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{gene}{NCBI gene ID}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
An object of class \code{DataGeNET.RDF}
}
\description{
Query for given gene(s) and generates an \code{DataGeNET.RDF}
Given the NCBI gene identifier, retrieves the information
related to drugs which interacts with the gene product and
their indications and creates an object of type \code{DataGeNET.RDF}.
}
