% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_disgenet_api_key.R
\name{get_disgenet_api_key}
\alias{get_disgenet_api_key}
\title{Retrieves the DisGeNET API key for a user}
\usage{
get_disgenet_api_key(email, password, verbose = FALSE, warnings = TRUE)
}
\arguments{
\item{email}{the email of the user}

\item{password}{the password of the user to connect to DisGeNET}

\item{verbose}{By default \code{FALSE}. Change it to \code{TRUE} to get a
on-time log from the function.}

\item{warnings}{By default \code{TRUE}. Change it to \code{FALSE} to don't see
the warnings.}
}
\value{
A string
}
\description{
Given the email and password, retrieves the users API Key
}
\examples{
disgenet_api_key <- get_disgenet_api_key( email = "user@gmail.com", password = "myspwd" )
}
