% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_disgenet_version.R
\name{get_disgenet_version}
\alias{get_disgenet_version}
\title{Retrieves DisGeNET data version}
\usage{
get_disgenet_version()
}
\description{
Retrieves DisGeNET data version
}
\examples{
dgn_version <- get_disgenet_version()
}
