% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeNET_DGN-ndisease.R
\name{ndisease}
\alias{ndisease}
\alias{DataGeNET.DGN-methods}
\title{Obtain the number of unique diseases in a \code{DataGeNET.DGN}.}
\usage{
\S4method{ndisease}{DataGeNET.DGN}(object)
}
\arguments{
\item{object}{Object of class \code{DataGeNET.DGN}}
}
\value{
The number of unique diseases
}
\description{
Obtain the number of unique diseases in a \code{DataGeNET.DGN}.
}
\examples{
\dontrun{
#Being x an DataGeNET.DGN
nd <- ndisease(x) # Get number of unique diseases
}
}
