% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeNET_DGN-ngene.R
\name{ngene}
\alias{ngene}
\alias{DataGeNET.DGN-methods}
\title{Obtain the number of unique genes in a \code{DataGeNET.DGN}.}
\usage{
\S4method{ngene}{DataGeNET.DGN}(object)
}
\arguments{
\item{object}{Object of class \code{DataGeNET.DGN}}
}
\value{
The number of unique genes
}
\description{
Obtain the number of unique genes in a \code{DataGeNET.DGN}.
}
\examples{
\dontrun{
#Being x an DataGeNET.DGN
nd <- ngene(x) # Get number of unique genes
}
}
