% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeNET-plot.R, R/DisGeNET-plot.R
\name{plot}
\alias{plot}
\alias{DataGeNET.DGN-methods}
\title{Plots the content of a \code{DataGeNET.DGN} object.
This functions allows to create a variety of plots for \code{DataGeNET.DGN} objects.}
\usage{
\S4method{plot}{DataGeNET.DGN}(x, y, ...)

\S4method{plot}{DataGeNET.DGN}(x, y, ...)
}
\arguments{
\item{x}{Object of class \code{DataGeNET.DGN}}

\item{y}{NOT USED}

\item{...}{Passed to inner functions for different plots.}

\item{layout}{Function to design the location of the different nodes. By
default \code{layout.fruchterman.reingold} from \code{igraph} is used.}

\item{class}{Type of the drawn chart. By default it is \code{"network"} but
it also can be \code{"Piechart"}, \code{"DiseaseClass"}, \code{"Venn"}
\code{"Heatmap"}, \code{"ProteinClass"}, \code{"Barplot"}, or \code{"Lollipop"}.}

\item{verbose}{By default \code{FALSE}. If set to \code{TRUE} information
on the drawing process will be shown.}
}
\value{
A plot for \code{DataGeNET.DGN}.

A plot for \code{DataGeNET.DGN}.
}
\description{
Plots the content of a \code{DataGeNET.DGN} object.
This functions allows to create a variety of plots for \code{DataGeNET.DGN} objects.

Plots the content of a \code{DataGeNET.DGN} object.
This functions allows to create a variety of plots for \code{DataGeNET.DGN} objects.
}
\examples{
\dontrun{
#Being x an DataGeNET.DGN
qr <- plot(x) # Get number of unique diseases
}
\dontrun{
Being x an DataGeNET.DGN
qr <- plot(x) # Get number of unique diseases
}
}
